celery fixes, add urls for approval and declining
