+++ investors/views.py
@@ -4,8 +4,6 @@ from rest_framework import permissions, status, viewsets
 from rest_framework.decorators import action
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
-from users.models import CustomUser
-
 
 from projects.models import Project
 from projects.serializers import ProjectSerializer
