+++ notifications/urls.py
@@ -1,4 +1,9 @@
 from django.urls import path
+from .views import approve, decline
 
+urlpatterns = [
+    path("approve/", approve, name='approve'),
+    path("decline/", decline, name='decline'),
+]
